Refactor engines into classes
