+++ CourageScores/ClientApp/src/components/tournaments/layout/ILayoutEngine.ts
@@ -1,0 +1,6 @@
+import {TournamentSideDto} from "../../../interfaces/models/dtos/Game/TournamentSideDto";
+import {ILayoutDataForRound} from "../layout";
+
+export interface ILayoutEngine {
+    calculate(sides: TournamentSideDto[]): ILayoutDataForRound[]
+}
